alteracao dos services e controller

+++ src/main/java/br/com/tecnoDesk/TecnoDesk/Services/ServicoItemService.java
@@ -1,10 +1,18 @@
 package br.com.tecnoDesk.TecnoDesk.Services;
 
 import br.com.tecnoDesk.TecnoDesk.DTO.ServicoItemDTO;
+import br.com.tecnoDesk.TecnoDesk.Entities.Empresa;
 import br.com.tecnoDesk.TecnoDesk.Entities.ServicoItem;
+import br.com.tecnoDesk.TecnoDesk.Repository.EmpresaRepository;
 import br.com.tecnoDesk.TecnoDesk.Repository.ServicoItemRepository;
 import lombok.RequiredArgsConstructor;
+import org.apache.coyote.BadRequestException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -12,33 +20,61 @@ public class ServicoItemService {
 
 	private final ServicoItemRepository servicoItemRepository;
 
-	public ServicoItem criarServicoItem(ServicoItemDTO servicoItemDTO) {
-		ServicoItem servicoItem = new ServicoItem();
-		servicoItem.setNome_servico(servicoItemDTO.nomeServico());
-		servicoItem.setDesc_servico(servicoItemDTO.descServico());
-		servicoItem.setValor_servico(servicoItemDTO.valorServico());
-		servicoItem.setCusto_servico(servicoItemDTO.custoServico());
-		return servicoItemRepository.save(servicoItem);
+	@Autowired
+	DecriptService decriptService;
+
+	@Autowired
+	EmpresaRepository empresaRepository;
+
+	public ServicoItem criarServicoItem(ServicoItemDTO servicoItemDTO, String codEmpresa) throws BadRequestException {
+		try {
+			Empresa empresa = empresaRepository.findEmpresaById(Long.valueOf(decriptService.decriptCodEmp(codEmpresa)));
+			if (empresa == null) {
+				throw new BadRequestException("Empresa não encontrada para o código fornecido.");
+			}
+			ServicoItem servicoItem = new ServicoItem();
+			servicoItem.setNomeServico(servicoItemDTO.nomeServico());
+			servicoItem.setDescServico(servicoItemDTO.descServico());
+			servicoItem.setValorServico(servicoItemDTO.valorServico());
+			servicoItem.setCustoServico(servicoItemDTO.custoServico());
+			servicoItem.setEmpresa(empresa);
+			return servicoItemRepository.save(servicoItem);
+
+		} catch (Exception e) {
+			throw new BadRequestException("Não foi possível criar um Servico, Por favor verifique os campos Nome do Servico", e);
+		}
 	}
+
+	public List<ServicoItem> buscarTodos() {
+		return servicoItemRepository.findAll();
+	}
+
+
 
 	public ServicoItemDTO atualizarServicoItem(Long id, ServicoItemDTO servicoItemDTO) {
 		return servicoItemRepository.findById(id)
 				.map(servicoItem -> {
-					servicoItem.setNome_servico(servicoItemDTO.nomeServico());
-					servicoItem.setDesc_servico(servicoItemDTO.descServico());
-					servicoItem.setValor_servico(servicoItemDTO.valorServico());
-					servicoItem.setCusto_servico(servicoItemDTO.custoServico());
+					servicoItem.setNomeServico(servicoItemDTO.nomeServico());
+					servicoItem.setDescServico(servicoItemDTO.descServico());
+					servicoItem.setValorServico(servicoItemDTO.valorServico());
+					servicoItem.setCustoServico(servicoItemDTO.custoServico());
 					ServicoItem atualizado = servicoItemRepository.save(servicoItem);
 					return new ServicoItemDTO(
-							atualizado.getNome_servico(),
-							atualizado.getDesc_servico(),
-							atualizado.getValor_servico(),
-							atualizado.getCusto_servico());
+							atualizado.getNomeServico(),
+							atualizado.getDescServico(),
+							atualizado.getValorServico(),
+							atualizado.getCustoServico());
 				})
 				.orElseThrow(() -> new RuntimeException("Serviço não encontrado"));
 	}
 
-	public void deletarServicoItem(Long id) {
-		servicoItemRepository.deleteById(id);
+	public ResponseEntity<String> deletarServicoItem(Long id) {
+		if (servicoItemRepository.existsById(id)) {
+			servicoItemRepository.deleteById(id);
+			return ResponseEntity.ok("Serviço com ID " + id + " foi deletado com sucesso.");
+		} else {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Serviço com ID " + id + " não encontrado.");
+		}
 	}
+
 }
